requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7